Moved methods that had to do with passport connectivity to a new class
