+++ app/src/main/AndroidManifest.xml
@@ -19,27 +19,13 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".jMRTDActivity"
-            android:label="jMRTDActivity" >
+            android:name=".PassportConActivity"
+            android:label="PassportConActivity" >
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
-  <!--          <intent-filter>
-                <action android:name="android.nfc.action.NDEF_DISCOVERED" />
-                <category android:name="android.intent.category.DEFAULT" />
-                <data android:mimeType="application/com.lgandroid.ddcnfc" />
-            </intent-filter>
-
-            <intent-filter>
-                <action android:name="android.nfc.action.TECH_DISCOVERED" />
-            </intent-filter>
-
-            <meta-data
-                android:name="android.nfc.action.TECH_DISCOVERED"
-                android:resource="@xml/nfc_tech_filter" />-->
         </activity>
 
     </application>
