+++ app/src/main/java/com/digitalvotingpass/digitalvotingpass/PassportConActivity.java
@@ -7,22 +7,14 @@ import android.content.Intent;
 import android.content.IntentFilter;
 import android.nfc.NfcAdapter;
 import android.nfc.Tag;
-import android.nfc.tech.IsoDep;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.widget.TextView;
 import android.widget.Toast;
 
-import net.sf.scuba.smartcards.CardService;
-
-import org.jmrtd.BACKeySpec;
 import org.jmrtd.PassportService;
-import org.jmrtd.lds.DG15File;
-import org.jmrtd.lds.DG1File;
-import org.jmrtd.lds.LDSFileUtil;
 import org.spongycastle.jce.provider.BouncyCastleProvider;
 
-import java.io.InputStream;
 import java.security.PublicKey;
 import java.security.Security;
 
@@ -44,7 +36,7 @@ public class PassportConActivity extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         Intent intent = getIntent();
 
-        setContentView(R.layout.activity_j_mrtd);
+        setContentView(R.layout.activity_passport_con);
         TextView textView = (TextView) findViewById(R.id.textView);
 
         mNfcAdapter = NfcAdapter.getDefaultAdapter(this);
@@ -131,8 +123,7 @@ public class PassportConActivity extends AppCompatActivity {
      * Handle the intent following from a NFC detection.
      *
      */
-    private void handleIntent(Intent intent){
-        TextView textView = (TextView) findViewById(R.id.textView);
+    private void handleIntent(Intent intent) {
         TextView textSignedData = (TextView) findViewById(R.id.signedData);
 
         // if nfc tag holds no data, return
@@ -142,23 +133,18 @@ public class PassportConActivity extends AppCompatActivity {
         }
 
         // Open a connection with the ID, return a PassportService object which holds the open connection
-        PassportService ps = openConnection(tag);
-        InputStream is = null;
-        InputStream is15 = null;
+        PassportConnection pcon= new PassportConnection();
+        PassportService ps = pcon.openConnection(tag);
         try {
             // Get the BSN from datagroup1 to confirm the ID was scanned correctly.
             // This is for testing purposes
-            is = ps.getInputStream(PassportService.EF_DG1);
-            DG1File dg1 = (DG1File) LDSFileUtil.getLDSFile(PassportService.EF_DG1, is);
-            Toast.makeText(this, dg1.getMRZInfo().getPersonalNumber(), Toast.LENGTH_LONG).show();
+            Toast.makeText(this, pcon.getBSN(ps), Toast.LENGTH_LONG).show();
 
             // display data from dg15
-            is15 = ps.getInputStream(PassportService.EF_DG15);
-
-            PublicKey pubk = getAAPublicKey(ps,is15);
+            PublicKey pubk = pcon.getAAPublicKey(ps);
 
             // sign 8 bytes of data and display the signed data
-            textSignedData.setText(byteArrayToHexString(signData(ps,is15,pubk)));
+            textSignedData.setText(PassportConnection.byteArrayToHexString(pcon.signData(ps, pubk)));
 
         } catch (Exception ex) {
             ex.printStackTrace();
@@ -166,122 +152,10 @@ public class PassportConActivity extends AppCompatActivity {
         } finally {
             try {
                 ps.close();
-                is.close();
-                is15.close();
             } catch (Exception e) {
                 e.printStackTrace();
                 Toast.makeText(this, R.string.general_error, Toast.LENGTH_LONG).show();
             }
         }
-    }
-
-    /**
-     * Opens a connection with the ID by doing BAC
-     * Uses hardcoded parameters for now
-     * @param tag - NFC tag that started this activity (ID NFC tag)
-     * @return PassportService - passportservice that has an open connection with the ID
-     */
-    private PassportService openConnection(Tag tag) {
-        PassportService ps = null;
-        try {
-            IsoDep nfc = IsoDep.get(tag);
-            CardService cs = CardService.getInstance(nfc);
-            ps = new PassportService(cs);
-            ps.open();
-
-            // Get the information needed for BAC, hardcoded for now
-            // TODO: link this with OCR functionality
-            ps.sendSelectApplet(false);
-            BACKeySpec bacKey = new BACKeySpec() {
-                @Override
-                public String getDocumentNumber() {
-                    return "NP0811B03";
-                }
-                @Override
-                public String getDateOfBirth() {
-                    return "940610";
-                }
-                @Override
-                public String getDateOfExpiry() {
-                    return "180624";
-                }
-            };
-
-            ps.doBAC(bacKey);
-            return ps;
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            ps.close();
-            Toast.makeText(this, R.string.general_error, Toast.LENGTH_LONG).show();
-        }
-        return null;
-    }
-
-
-    /**
-     * Retrieves the public key used for Active Authentication from datagroup 15.
-     * @return Publickey - returns the publickey used for AA
-     */
-    private PublicKey getAAPublicKey(PassportService ps, InputStream is15) {
-        try {
-            DG15File dg15 = new DG15File(is15);
-            return dg15.getPublicKey();
-        } catch (Exception ex){
-            ex.printStackTrace();
-            Toast.makeText(this, R.string.general_error, Toast.LENGTH_LONG).show();
-        }
-        return null;
-    }
-
-    /**
-     * Signs 8 bytes by the passport using the AA functionality.
-     * @return byte[] - signed byte array
-     */
-    private byte[] signData(PassportService ps, InputStream is15, PublicKey pubk) {
-        try {
-            // test 8 byte string for testing purposes
-            byte[] data = hexStringToByteArray("0a1b3c4d5e6faabb");
-            return ps.doAA(pubk, null, null, data);
-        } catch (Exception ex) {
-            ex.printStackTrace();
-            Toast.makeText(this, R.string.general_error, Toast.LENGTH_LONG).show();
-        }
-        return null;
-    }
-
-    /**
-     * Method for converting a hexString to a byte array.
-     * This method is used for signing transaction hashes (which are in hex).
-     */
-    public static byte[] hexStringToByteArray(String hStr) {
-        if(hStr != null) {
-            int len = hStr.length();
-            byte[] data = new byte[len / 2];
-            for (int i = 0; i < len; i += 2) {
-                data[i / 2] = (byte) ((Character.digit(hStr.charAt(i), 16) << 4)
-                        + Character.digit(hStr.charAt(i + 1), 16));
-            }
-            return data;
-        }
-        return new byte[0];
-    }
-
-    /**
-     * Method for converting a byte array to a hexString.
-     * This method is used for converting a signed 8-byte array back to a hashString in order to
-     * display it readable.
-     */
-    public static String byteArrayToHexString(byte[] bArray) {
-        if (bArray != null) {
-            final char[] hexArray = "0123456789ABCDEF".toCharArray();
-            char[] hexChars = new char[bArray.length * 2];
-            for (int j = 0; j < bArray.length; j++) {
-                int v = bArray[j] & 0xFF;
-                hexChars[j * 2] = hexArray[v >>> 4];
-                hexChars[j * 2 + 1] = hexArray[v & 0x0F];
-            }
-            return new String(hexChars);
-        }
-        return "";
     }
 }
