+++ app/src/main/res/layout/activity_passport_con.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.digitalvotingpass.digitalvotingpass.jMRTDActivity">
+    tools:context="com.digitalvotingpass.digitalvotingpass.PassportConActivity">
 
     <TextView
         android:id="@+id/textView"
